Phase 2 audit follow-ups: split useDashboardStats + WidgetCatalog; correct README
Three audit items from the Phase 2 self-review. With this commit no source
file under src/admin-app/src/ is over 300 lines and the README no longer
contradicts what the app actually does.

useDashboardStats decomposition (393 -> 177 lines)
- Move 13 pure aggregators (buildCategoryData, buildChallengeSummary,
  buildComparisonData, buildStatusBreakdown, buildUserGrowth,
  buildGroupPerformance, buildCompletionRates, buildPointsDistribution,
  buildMostActiveUsers, buildRecentActivity, plus truncate / sumPoints
  / countUsersCreatedThisMonth) into hooks/aggregations.js.
- Hook owns orchestration only (useEffect + Promise.all + buildStats).
- Drops the previous "splitting would hide the helpers" justification
  comment; helpers are now exported and independently unit-testable.

WidgetCatalog decomposition (304 -> 88 lines)
- The in-file section dividers were already extraction markers. Split
  into components/catalog/{CatalogPresets,CatalogChallengeFilter,
  CatalogWidgetList}.jsx (39 / 120 / 136 lines respectively).
- Parent now composes drawer chrome + the four section components +
  the reset footer; no other behavior change.

README correctness pass
- Replace "Login page with mock authentication" with the real Supabase
  magic-link flow; remove the demo-account password table.
- Replace "localStorage persistence" / "Reset Demo Data" claims with
  the truth (Supabase is the source of truth; localStorage caches
  only the per-user dashboard layout).
- Drop the obsolete "Planned: FastAPI / standalone PostgreSQL /
  OAuth-JWT" tech-stack table -- Supabase replaced all three in v0.6.x.
- Expand Quick Start with the .env step (was implicit before).
- Rewrite "Future Goals" to drop the shipped items (FastAPI / Postgres
  / auth) and add the actually-open ones (shared schema with Team 1,
  code-splitting to fix the 1.3 MB bundle warning, test suite).
- Bump version label to v0.7.2; add a changelog entry covering this
  commit.

Verified: npm run lint (0 errors, 0 warnings -- the previous
useDashboardStats max-lines warning is gone), npm run format:check
(clean), npm run build (passes; bundle-size warning is the pre-existing
one already tracked in Future Goals).
